requirements.txt
Pillow==9.4.0
qrcode==7.4.2
streamlit==1.19.0
